feature/DTO Added lombok to the project, added README.md file

+++ src/main/java/com/md/demo/repository/RatingRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RatingRepository extends CrudRepository<Rating, Long> {
+public interface RatingRepository extends CrudRepository<Rating, Integer> {
 
 }
